Update color-maps, prevent bleed
Rather than using a direct math relation-ship
  `color-index-number = MIDI-note - 36`
this allow to map MIDI notes arbitrarily to colors.
This helps when scaling the GUI and map colors are
interpolated between neighbors, and fixes a bug where those
a region between Hi-Hat swish and open hat triggered a tom hit.

It also allows to constrain the visual highlight to a given
kit piece.